Refactor get_time_stamp_ns
