+++ src/mono/mono/mini/mini-runtime.c
@@ -70,6 +70,7 @@
 #include <mono/utils/mono-compiler.h>
 #include <mono/utils/mono-proclib.h>
 #include <mono/utils/mono-state.h>
+#include <mono/utils/mono-time.h>
 #include <mono/metadata/w32handle.h>
 #include <mono/metadata/threadpool.h>
 
@@ -1980,7 +1981,6 @@ typedef struct
 } JitCodeLoadRecord;
 
 static void add_file_header_info (FileHeader *header);
-static guint64 get_time_stamp_ns (void);
 static void add_basic_JitCodeLoadRecord_info (JitCodeLoadRecord *record);
 
 void
@@ -2021,16 +2021,8 @@ add_file_header_info (FileHeader *header)
 	header->elf_mach = ELF_MACHINE;
 	header->pad1 = 0;
 	header->pid = perf_dump_pid;
-	header->timestamp = get_time_stamp_ns ();
+	header->timestamp = mono_clock_get_time_ns ();
 	header->flags = 0;
-}
-
-static guint64
-get_time_stamp_ns (void)
-{
-	struct timespec ts;
-	int result = clock_gettime (CLOCK_MONOTONIC, &ts);
-	return  ts.tv_sec * 1000000000ULL + ts.tv_nsec;
 }
 
 void
@@ -2055,7 +2047,7 @@ mono_emit_jit_dump (MonoJitInfo *jinfo, gpointer code)
 		
 		// TODO: write debugInfo and unwindInfo immediately before the JitCodeLoadRecord (while lock is held).
 		
-		record.header.timestamp = get_time_stamp_ns ();
+		record.header.timestamp = mono_clock_get_time_ns ();
 		
 		fwrite (&record, sizeof (record), 1, perf_dump_file);
 		fwrite (jinfo->d.method->name, nameLen + 1, 1, perf_dump_file);
@@ -2069,7 +2061,7 @@ static void
 add_basic_JitCodeLoadRecord_info (JitCodeLoadRecord *record)
 {
 	record->header.id = JIT_CODE_LOAD;
-	record->header.timestamp = get_time_stamp_ns ();
+	record->header.timestamp = mono_clock_get_time_ns ();
 	record->pid = perf_dump_pid;
 	record->tid = syscall (SYS_gettid);
 }
